DependencyGraph: use UpcastableType in TypeNode
This is the foundation to start supporting TypeNodes representing
artificial struct wrappers in the DependencyGraph.
Given that a type wrapped in an array may not be a
model::TypeDefinition, we need the full expressive power of
UpcastableType.

At the moment, no TypeNode representing an artificial struct wrapper is
created yet.