TypeNames: reduce wrappers for pointers to array
With this commit, artificial struct wrappers are now injected only for
pointers to arrays whose fields are structs, unions, or typedef of
struct or unions.

The change is motivated by the fact that always adding artificial struct
wrappers around arrays that are pointed to by pointers is too heavy, and
the generated wrappers are very ugly and verbose.

To enable this reduction of the number of wrappers, CTypeBuilder had to
be updated to not only track what are the types that should be inlined,
but also *where* they should be inlined.