Fix revng pipe output container targets
Previously revng pipe did not correctly propagate targets from input
containers into output containers based on the pipe contract. This
happened to work for pipes acting on LLVM containers where the same
container is used for input and output, but failed for the Clift backend
pipe, where an MLIR container is used for input and a string map
container is used for output.