Re-enable tests formerly disabled do to a bug
This commit turns on some old tests that were disabled for a bug.

The comment on the disabled tests said that it was due to a bug in type
inlining, but it was actually caused by a bug in the computation of
dependencies in DepenencyGraph.

Type inlining is now gone, and DependencyGraph has been fixed, so it's
now possible to re-enable these tests.