CSAA: fix return value on changes
Before this commit, `forceEmptyMetadata` did not return a bool to notify
when it changed something.
This in turn caused the pass to return the wrong value when the only
change applied was to force the empty metadata.

This commit fixes the issue, letting `forceEmptyMetadata` return a bool,
and propagating that until where it matters.