Pypeline: StorageProviderFactory
Now each storage provider has its own factory so we can cache instances
on local development and reuse connections in the cloud version.
As a consequence of this refactor, the user can add its own storage
solution by implementing a StorageProvider and StorageProviderFactory.
Each factory has a `protocol` static method that will be used to
initialize the correct factory based on the user-provided
storage-provider-url.