AVI: sort entry points by dominance
This commit a non-deterministic bug on ARM's switch-disjoint-ranges
test.

The problem was that we were setting as initial nodes of the monotone
framework all the nodes that are not reachable from the nodes classified
as initial nodes *so far*.
This approach is susceptible to the visit order.
Classifying certain nodes as entry nodes degrades the analysis results
since we associates a top value to them.

Electing initial nodes by traversing a list of nodes sorted by dominance
solves this problem.