LLVMContainer: fix metadata handling
The way we were saving metadata during cloneFiltered was incorrect and
led to subtle corruptions due to reattaching metadata coming from one
module to the other.

We now collect all the metadata we want to preserve in a named metadata
(a "global" metadata), so that it will get properly copied.